add: multi-hostname support
flake.nix now has the ability to not only define multiple nixos systems
to build from, but the ability to change the hostname for each one
without changing the actual contents. this is all i'm going to do for
now, until i require different system and user configurations for each
host.